Added query options module. #2
