+++ public/index.html
@@ -8,9 +8,9 @@
   </head>
   <body>
     <div class="container">
-      <form class="form-inline">
-        <input type="text" class="form-control" size="40" placeholder="Job title, keyword, or company"/>
-        <input id="search-text-field" type="text" class="form-control" size="30" placeholder="Enter a city"/>
+      <form id="job-search" class="form-inline">
+        <input id="job-input" type="text" class="form-control" size="40" placeholder="Job title, keyword, or company"/>
+        <input id="city-input" type="text" class="form-control" size="30" placeholder="Enter a city"/>
         <button type="submit" class="btn btn-default">Submit</button>
       </form>
     </div>
@@ -24,7 +24,7 @@
     <script>
       function initialize() {
         const options = { types: ['(cities)'] }
-        const input = document.querySelector('#search-text-field')
+        const input = document.querySelector('#city-input')
         const autocomplete = new google.maps.places.Autocomplete(input, options)
       }
     </script>
