+++ query-options.js
@@ -1,0 +1,10 @@
+const queryOptions = {
+  query: 'JavaScript',
+  city: 'Irvine, CA',
+  radius: '25',
+  level: 'entry_level',
+  jobType: 'fulltime',
+  sort: 'date',
+  limit: '25'
+}
+module.exports = queryOptions
